Added session_id and window_id to ViewInstance
This is to be able to track which session and window a view plugin belongs to. I'm not really sure about this design but allows the code to only have one owner of the plugins and they just gets sent down to the other system without them actually needing to store them.