Added support for text_fmt(...)
It uses "format_args!" to format a string that can be used for printing text. In order to reduce temporary allocations Ui instance now keeps a string around that it uses when constructing the data instead of creating a new one all the time.

Because of this change Ui must now be sent as mutable as the write!(...) macro will muteate the temporary string.